cache: make similar to sync.Map
Internally, read heavy workloads now avoid a map. Since the point of a
cache is to be read heavy, this should help the workloads that this
struct is intended for.

This mirrors some logic of sync.Map, but makes different decisions on
when to promote and has to account for a slow load. We also avoid
mirroring the entire read map into dirty up front, and instead slow down
the promotion path.